adds back columns and row code
