+++ apps/backend/src/config/dependencyConfigSTFC.ts
@@ -35,8 +35,8 @@ import {
   createSkipListeningHandler,
 } from '../eventHandlers/messageBroker';
 import { createApplicationEventBus } from '../events';
+import { StfcSEPDataColumns } from '../factory/xlsx/SEPDataColumns';
 import { getStfcDataRow } from '../factory/xlsx/SEPDataRow';
-import { StfcSEPDataColumns } from '../factory/xlsx/StfcSEPDataColumns';
 import { SkipAssetRegistrar } from '../services/assetRegistrar/skip/SkipAssetRegistrar';
 import { configureSTFCEnvironment } from './stfc/configureSTFCEnvironment';
 import { Tokens } from './Tokens';
