+++ apps/backend/src/factory/xlsx/SEPDataColumns.ts
@@ -9,3 +9,17 @@ export const SEPDataColumns = [
   'Current rank',
   'Is in availability zone',
 ];
+
+export const StfcSEPDataColumns = [
+  'Proposal Reference Number',
+  'Proposal Title',
+  'Principal Investigator',
+  'Instrument Name',
+  'Instrument Available Time',
+  'Technical Review Allocated Time',
+  'FAP Allocated Time',
+  'Average Score',
+  'Current Rank',
+  'Is in Availability Zone',
+  'Feedback',
+];
