--- apps/backend/src/factory/xlsx/StfcSEPDataColumns.ts
@@ -1,13 +1,0 @@
-export const StfcSEPDataColumns = [
-  'Proposal Reference Number',
-  'Proposal Title',
-  'Principal Investigator',
-  'Instrument Name',
-  'Instrument Available Time',
-  'Technical Review Allocated Time',
-  'FAP Allocated Time',
-  'Average Score',
-  'Current Rank',
-  'Is in Availability Zone',
-  'Feedback',
-];
